+++ client/assets/main.styl
@@ -4,6 +4,7 @@
 @import '../components/admin_form'
 @import '../components/article_list'
 @import '../components/autocomplete'
+@import '../components/autocomplete2'
 @import '../components/autocomplete_select'
 @import '../components/character_limit'
 @import '../components/drag_drop'
