+++ client/components/autocomplete/index.styl
@@ -105,54 +105,3 @@
   line-height 1em
   padding 2px 20px 1px 4px
   text-transform capitalize
-
-
-.Autocomplete
-  position relative
-  z-index 10
-  input
-    padding-left 36px
-    width 100%
-    &[disabled]
-      background-color gray-lightest-color !important
-  &__icon
-    position absolute
-    top 7px
-    left 3px
-  &__results
-    border-left 1px solid gray-lighter-color
-    border-right 1px solid gray-lighter-color
-    position absolute
-    z-index 10
-  &__results-bg
-      position fixed
-      left 0
-      right 0
-      top 0
-      bottom 0
-      opacity 0
-      z-index -1
-  &__item
-    padding 10px
-    background white
-    color black
-    display flex
-    align-items center
-    garamond l-body
-    border-bottom 1px solid gray-lighter-color
-    z-index 3
-    &:hover
-      background gray-lightest-color
-      cursor pointer
-    &---empty
-      background gray-lighter-color
-      color gray-darker-color
-  &__item-img
-    width 45px
-    height 45px
-    margin-right 15px
-    background gray-lighter-color
-    img
-      height 45px
-      width 45px
-      object-fit cover
